# Add CSE to aie-opt pipeline to improve -O2 performance