This PR extends the aie-opt pipeline with Common Subexpression Elimination (CSE) to achieve better scalar arithmetic without increasing Program Memory (PM) usage.

Background:
- Currently, -O3 optimization provides better performance than -O2  (function outlining, phoenix, and matmul). 
- However, -O3 significantly increases program memory usage compared to -O2.
- The performance improvement in -O3 is largely due to unintended effects of function inlining, which rearranges integer arithmetic and reduces common subexpressions.

Problem:
- In -O2, the outlined function after optimization contains redundant operations, e.g.:

%45 = trunc i64 %38 to i20
...
%49 = trunc i64 %38 to i20

- -O3 eliminates these redundancies indirectly through inlining, but at the cost of increased memory usage.

Solution:
- Add a basic LLVM CSE pass after the -O* pass pipeline.
- This approach achieves most/all of the performance benefits of -O3 without the memory overhead.

Implementation:
- Extended aie-opt pipeline with CSE.
- Considered adding Global Value Numbering (GVN) but found it merged vector loads, causing register spills and performance degradation.

References:
- Early CSE information: https://llvm.org/doxygen/structllvm_1_1EarlyCSEPass.html#details

Below I've scraped all the benchmark numbers from CI, for this PR and top of main (ToM). Table summary: it looks like we can match the performance of -O3 (line 4) now with -O2 (line 2), but with much less PM usage (8474 bytes -> 2608 bytes). 

### Time Comparison
| Benchmark | Before (us) | After (us) |
| --- | --- | --- |
| matmul_512_512_4096_bf16_f32_O2_npu1_4col | 2585.0 | 1817.0 |
| matmul_512_512_4096_bf16_f32_O2_npu1_4col_outline | 2714.0 | **1807.0** |
| matmul_512_512_4096_bf16_f32_O3_npu1_4col | 2625.0 | 1801.0 |
| matmul_512_512_4096_bf16_f32_O3_npu1_4col_outline | **1775.0** | 1804.0 |
| matmul_512_512_4096_bf16_f32_O3_npu1_4col_outline_ukernel_chess | 1771.0 | 1832.0 |


### Memory Comparison
| Benchmark | Before (bytes) | After (bytes) |
| --- | --- | --- |
| matmul_512_512_4096_bf16_f32_O2_npu1_4col | 10112 | 8800 |
| matmul_512_512_4096_bf16_f32_O2_npu1_4col_outline | 2784 | **2608** |
| matmul_512_512_4096_bf16_f32_O3_npu1_4col | 10112 | 8800 |
| matmul_512_512_4096_bf16_f32_O3_npu1_4col_outline | **8464** | 8464 |
| matmul_512_512_4096_bf16_f32_O3_npu1_4col_outline_ukernel_chess | 3328 | 3328 |