@yzhang93 2025-02-27
One more request: could you truncate the table when you squash and merge the PR?