@newling 2025-02-27
> One more request: could you truncate the table when you squash and merge the PR?

Sure, but I'm not sure why? Are you worried it'll look bad when you do `git log`? Because if so it's not as bad as PRs like https://github.com/nod-ai/iree-amd-aie/pull/1115 which look really untidy in terminal