@newling 2025-02-27
> This is great! Have you tried to compare O2 and O3 results using other shapes?

I haven't, no