@yzhang93 2025-02-27
> > One more request: could you truncate the table when you squash and merge the PR?
> 
> Sure, but I'm not sure why? Are you worried it'll look bad when you do `git log`? Because if so it's not as bad as PRs like #1115 which look really untidy in terminal

Yes, the log would be quite long. Besides I think the first 5 rows are enough to demonstrate your contribution.